Add very rough implementation of APU
With this change, most games' audio is at least recognizable.  However,
it sounds pretty terrible.  At the moment (on my machine), the emulator
runs below 60 FPS, which leaves SDL_QueueAudio with noticeable periods
of silence on every frame.  For debugging the APU, this makes it very
difficult to notice any actual glitches in the audio channels' output.
This system should be refactored after the emulator has been optimized
to consistently run at 60 FPS.