Simplify DIV resets and make gb->cycles persistent
When the DIV register is reset, it appears that only the APU frame
sequencer and Timer are affected as they are clocked on particular bit
transitions (see: TCAGBD).  This means we don't need to reset
gb->cycles; we just need to keep track of the last time DIV was reset.

In addition, this also allows systems outside of the core GB to use
cycles as a reliable counter, which could be useful for an external link
cable implementation (coming soon™), or more consistent joypad sampling.