Improve GB synchronization in local serial mode
Now that gb->cycles is persistent, we can use it as a much more reliable
synchronization tool than gameboy_tick(), simply due to the fact that
opcodes can take anywhere from 4 to 24 cycles to execute.  If the host's
code was CALL-heavy while the guest code was NOP-heavy, for example,
this could lead to data errors from SB being sampled at the wrong time.

The overall flow should now look something like this:

-	Run host for {max} cycles
-	Run guest for {max} cycles
-	Update joypad, repeat

-	Run host for {x} cycles (interrupted by start of serial xfer)
-	Run guest for {x} cycles
-	Start serial xfer (host and guest are both at {max + x} cycles)
-	Update joypad, repeat