Add plugin system to clean up curl/lws/ruby deps
This change refactors some of the ugly, highly-coupled code in egbe.h
and the Makefile.  The plugin API is still a work in progress, as are
the responsibilities of egbe_application and egbe_gameboy.

Notably, the init/exit/call hooks in the plugin API were necessary to
support the Ruby plugin.  To get the debugger to work properly, Ruby's
API requires all further library calls to exist ON THE SAME STACK FRAME,
which makes a simple init/exit API insufficient.

This plugin framework also makes it possible to extract SDL into its own
plugin, so that other libraries could be used for video/audio/keyboard.
Maybe later.